requirements.txt
lxml
requests
hitutil
ddddocr